--- lib/fsutils.ts
@@ -1,17 +1,0 @@
-import fs from "fs";
-import { resolve } from "path";
-
-/** Return a flat array of all files under given directory */
-export function getFilesRecursively(dir: string): string[] {
-  let found: string[] = [];
-  const dirents = fs.readdirSync(dir, { withFileTypes: true });
-  for (const dirent of dirents) {
-    const path = resolve(dir, dirent.name);
-    if (dirent.isDirectory()) {
-      found = found.concat(getFilesRecursively(path));
-    } else {
-      found.push(path);
-    }
-  }
-  return found;
-}
