+++ lib/post-loading.ts
@@ -1,6 +1,21 @@
 import fs from "fs";
+import { resolve } from "path";
 import { BlogPost, decodeBlogPost } from "./post";
-import { getFilesRecursively } from "./fsutils";
+
+/** Return a flat array of all files under given directory */
+export function getFilesRecursively(dir: string): string[] {
+  let found: string[] = [];
+  const dirents = fs.readdirSync(dir, { withFileTypes: true });
+  for (const dirent of dirents) {
+    const path = resolve(dir, dirent.name);
+    if (dirent.isDirectory()) {
+      found = found.concat(getFilesRecursively(path));
+    } else {
+      found.push(path);
+    }
+  }
+  return found;
+}
 
 /**
  * Read blog post from a file
